Add intervals

Since we have longitudinal data, intervals for subjects is a good way to get a better idea of the time scale of the data. These functions add extra columns to your data.

add_interval(data, name = interval)

add_interval_baseline(data, name = interval_baseline)

Arguments

data
data extracted from the NOAS

name
unquoted name to give the new column

Value

data frame with one extra column

Details

add_interval - add interval since last visit - default col: interval

add_interval_baaseline - add interval since first visit- default col: baseline

Examples

# attach built-in noas example data to test
dt <- noas_example

add_interval(dt)

#> # A tibble: 10 x 9
#>    subject_id project_id wave_code site_name mri_info_folder visit_age   cog
#>    <chr>      <chr>          <dbl> <chr>     <chr>               <dbl> <dbl>
#>  1 1000000    MemP               1 ousAvanto 1000000_1               8    16
#>  2 1000000    MemP               2 ousAvanto 1000000_2              10    14
#>  3 1000000    MemP               3 ousAvanto 1000000_3              14    16
#>  4 1000000    MemP               3 ousSkyra  1000000_4              14    NA
#>  5 1000000    MemP               4 ousSkyra  1000000_5              17    15
#>  6 1000000    MemP               5 ousSkyra  1000000_6              20    15
#>  7 1000010    MemC               1 ousSkyra  1000010_1              22    14
#>  8 1000010    MemC               2 ousSkyra  1000010_2              28    13
#>  9 1000010    MemC               3 ousSkyra  1000010_3              33    NA
#> 10 1000010    MemC               4 ousSkyra  1000010_4              40    10
#> # … with 2 more variables: sex <chr>, interval <dbl>

add_interval_baseline(dt)

#> # A tibble: 10 x 9
#>    subject_id project_id wave_code site_name mri_info_folder visit_age   cog
#>    <chr>      <chr>          <dbl> <chr>     <chr>               <dbl> <dbl>
#>  1 1000000    MemP               1 ousAvanto 1000000_1               8    16
#>  2 1000000    MemP               2 ousAvanto 1000000_2              10    14
#>  3 1000000    MemP               3 ousAvanto 1000000_3              14    16
#>  4 1000000    MemP               3 ousSkyra  1000000_4              14    NA
#>  5 1000000    MemP               4 ousSkyra  1000000_5              17    15
#>  6 1000000    MemP               5 ousSkyra  1000000_6              20    15
#>  7 1000010    MemC               1 ousSkyra  1000010_1              22    14
#>  8 1000010    MemC               2 ousSkyra  1000010_2              28    13
#>  9 1000010    MemC               3 ousSkyra  1000010_3              33    NA
#> 10 1000010    MemC               4 ousSkyra  1000010_4              40    10
#> # … with 2 more variables: sex <chr>, interval_baseline <dbl>

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dt %>%
  add_interval() %>%
  add_interval_baseline()

#> # A tibble: 10 x 10
#>    subject_id project_id wave_code site_name mri_info_folder visit_age   cog
#>    <chr>      <chr>          <dbl> <chr>     <chr>               <dbl> <dbl>
#>  1 1000000    MemP               1 ousAvanto 1000000_1               8    16
#>  2 1000000    MemP               2 ousAvanto 1000000_2              10    14
#>  3 1000000    MemP               3 ousAvanto 1000000_3              14    16
#>  4 1000000    MemP               3 ousSkyra  1000000_4              14    NA
#>  5 1000000    MemP               4 ousSkyra  1000000_5              17    15
#>  6 1000000    MemP               5 ousSkyra  1000000_6              20    15
#>  7 1000010    MemC               1 ousSkyra  1000010_1              22    14
#>  8 1000010    MemC               2 ousSkyra  1000010_2              28    13
#>  9 1000010    MemC               3 ousSkyra  1000010_3              33    NA
#> 10 1000010    MemC               4 ousSkyra  1000010_4              40    10
#> # … with 3 more variables: sex <chr>, interval <dbl>, interval_baseline <dbl>

# Change the default column names
dt %>%
  add_interval(name = intv) %>%
  add_interval_baseline(name = bsl_intv)

#> # A tibble: 10 x 10
#>    subject_id project_id wave_code site_name mri_info_folder visit_age   cog
#>    <chr>      <chr>          <dbl> <chr>     <chr>               <dbl> <dbl>
#>  1 1000000    MemP               1 ousAvanto 1000000_1               8    16
#>  2 1000000    MemP               2 ousAvanto 1000000_2              10    14
#>  3 1000000    MemP               3 ousAvanto 1000000_3              14    16
#>  4 1000000    MemP               3 ousSkyra  1000000_4              14    NA
#>  5 1000000    MemP               4 ousSkyra  1000000_5              17    15
#>  6 1000000    MemP               5 ousSkyra  1000000_6              20    15
#>  7 1000010    MemC               1 ousSkyra  1000010_1              22    14
#>  8 1000010    MemC               2 ousSkyra  1000010_2              28    13
#>  9 1000010    MemC               3 ousSkyra  1000010_3              33    NA
#> 10 1000010    MemC               4 ousSkyra  1000010_4              40    10
#> # … with 3 more variables: sex <chr>, intv <dbl>, bsl_intv <dbl>