Create a data.frame that is ready for use in Freesufers linear mixed models

fs_lmm Creates a data.frame of subsetted row and formatted columns ready for use in Freesurfers (6.0) linear mixed models. Design matrices are created for the grouping.var factors, and numeric variables are z-transformed.

fs_lmm(data, formula, site_var, folder_var, missing = "mean", ..., file = NULL)

Arguments

data
data extracted from the NOAS

formula
right-hand formula for your model. (ex. ~ age * sex)

site_var
column name with the site information

folder_var
unqoted column name with folder-information of MRI data from the 'mri_info' table.

missing
Action to take on missing data that are numeric

...
other arguments to utils{model.matrix}

file
Optional string specifying file name to be saved

Value

a data frame ready for Freesurfer LMM use.

Details

Available options for 'missing.action' are:

"delete"delete observations with any missing numeric.vars

"mean"replace missing with mean for that participant

"all"replace all values with the mean for that participant

"first"replace all values with the first observation for that participant

Examples

if (FALSE) {
fs_lmm(data,
       age * sex * cog
     )
}